+++ packages/number-input/src/Component.screenshots.test.tsx
@@ -5,11 +5,11 @@ describe('NumberInput ', () =>
         {
             componentName: 'NumberInput',
             knobs: {
-                label: 'Label',
+                label: 'Число',
                 value: 1234,
                 size: 'm',
                 block: true,
             },
         },
-        'padding: 0 270px; transform:scale(2.3)',
+        'padding: 0 270px; transform:scale(2.1)',
     ));