+++ packages/bottom-sheet/CHANGELOG.md
@@ -1,5 +1,23 @@
 # Change Log
 
+## 6.2.5
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   navigation-bar@0.3.3
+
+## 6.2.4
+
+### Patch Changes
+
+### [#654](https://github.com/core-ds/core-components/pull/654)
+
+-   Удалены лишние dependencies, добавлены отсутствующие
+
+-   Обновлены зависимости
+    -   navigation-bar@0.3.2
+
 ## 6.2.3
 
 ### Patch Changes