+++ packages/calendar-with-skeleton/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Change Log
 
+## 5.0.43
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   calendar@6.3.3
+
+## 5.0.42
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   calendar@6.3.2
+
 ## 5.0.41
 
 ### Patch Changes