+++ packages/calendar/CHANGELOG.md
@@ -1,5 +1,23 @@
 # Change Log
 
+## 6.3.3
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   button@8.3.0
+    -   icon-button@6.0.12
+    -   modal@7.0.10
+
+## 6.3.2
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   button@8.2.0
+    -   modal@7.0.9
+    -   icon-button@6.0.11
+
 ## 6.3.1
 
 ### Patch Changes