+++ packages/confirmation/CHANGELOG.md
@@ -1,5 +1,23 @@
 # Change Log
 
+## 12.4.15
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   button@8.3.0
+
+## 12.4.14
+
+### Patch Changes
+
+### [#654](https://github.com/core-ds/core-components/pull/654)
+
+-   Удалены лишние dependencies, добавлены отсутствующие
+
+-   Обновлены зависимости
+    -   button@8.2.0
+
 ## 12.4.13
 
 ### Patch Changes