+++ packages/date-time-input/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @alfalab/core-components-date-time-input
 
+## 3.1.8
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   calendar@6.3.3
+    -   icon-button@6.0.12
+    -   input@11.1.14
+
+## 3.1.7
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   calendar@6.3.2
+    -   icon-button@6.0.11
+    -   input@11.1.13
+
 ## 3.1.6
 
 ### Patch Changes