+++ packages/gallery/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Change Log
 
+## 5.2.6
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   icon-button@6.0.12
+    -   tooltip@6.2.21
+
+## 5.2.5
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   icon-button@6.0.11
+    -   tooltip@6.2.20
+
 ## 5.2.4
 
 ### Patch Changes