+++ packages/icon-button/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Change Log
 
+## 6.0.12
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   button@8.3.0
+
+## 6.0.11
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   button@8.2.0
+
 ## 6.0.10
 
 ### Patch Changes