+++ packages/pure-cell/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Change Log
 
+## 4.2.8
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   button@8.3.0
+
+## 4.2.7
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   button@8.2.0
+
 ## 4.2.6
 
 ### Patch Changes