+++ packages/system-message/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @alfalab/core-components-system-message
 
+## 1.0.2
+
+### Patch Changes
+
+### [#629](https://github.com/core-ds/core-components/pull/629)
+
+-   Исправлены заголовки в теме click
+
 ## 1.0.1
 
 ### Patch Changes